+++ app/Http/Controllers/LocationController.php
@@ -38,6 +38,11 @@ class LocationController extends Controller
 
     public function update(Request $request, Location $location)
     {
+        // Ako kategorija ima oglase, zabrani brisanje
+        if ($location->ads()->count() > 0) {
+            return redirect()->route('locations.index')->with('error', 'Ne možete urediti lokaciju jer ima povezane oglase.');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -51,7 +56,14 @@ class LocationController extends Controller
 
     public function destroy(Location $location)
     {
+        // Ako kategorija ima oglase, zabrani brisanje
+        if ($location->ads()->count() > 0) {
+            return redirect()->route('locations.index')->with('error', 'Ne možete obrisati lokaciju jer ima povezane oglase.');
+        }
+        // Sada možeš obrisati lokaciju
         $location->delete();
+
         return redirect()->route('locations.index')->with('success', 'Lokacija uspešno obrisana.');
     }
+
 }
